Construct Simultaneous Confidence Bands (SCB) For One Dimensional Functional Data Using Specified Methods

This function builds simultaneous confidence bands through two distinct approaches..

SCB_functional_outcome(
  data,
  object = NULL,
  method,
  fitted = TRUE,
  est_mean = TRUE,
  alpha = 0.05,
  outcome,
  time,
  range = NULL,
  group_name = NULL,
  group_value = NULL,
  subject,
  nboot = NULL,
  method_SD = "t",
  weights = "rademacher"
)

Arguments

data

A functional data frame, matrix or tibble. The input data must have column names, and should contain the functional outcome, time, subject

object

A fitted Function-on-Scalar Regression (FoSR) object (e.g., from mgcv::bam()/mgcv::gam()). Default is NULL

method

A character string specifying the approach:

"cma" - Correlation and Multiplicity Adjusted (CMA) confidence bands via parametric approach (requires a fitted functional regression model)

"multiplier" - Dense confidence bands via Multiplier-t Bootstrap method For method = "multiplier" option, the user should check whether their input data (data) has zero entries on every subject for a specific time point. The function only allow all zero entries for time point zero. Otherwise, the function will return an error. Besides, for Multiplier Bootstrap, if there exist NA in data, the function will impute these NA's via fpca.face.

fitted

Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function

TRUE - Estimate the simultaneous confidence bands for fitted mean outcome function.

FALSE - estimate the simultaneous confidence bands for fitted parameter function.

Default is TRUE.

est_mean

Logical. Whether to use the fitted mean from a functional regression model in Multiplier Bootstrap (method = "multiplier") for estimating the quantile. This argument can be ignored if choose method = "cma". If TRUE, the function will use:

Fitted mean: The estimated mean function from a fitted model (e.g., mgcv::bam). Need to provide a fitted functional outcome regression model object. Otherwise, the sample mean will be calculated.

If FALSE, the function will instead use:

Sample mean: $$\hat{\beta}(s) = \frac{1}{n} \sum_{i=1}^n \tilde{Y}_i(s)$$ where \(\tilde{Y}_i(s)\) is a smoothed version of the observed functional response. For this sample mean option, if there exists NA in data, will perform imputation to fill in those values.

Default is FALSE.

alpha

Significance level for SCB. Default is 0.05.

outcome

A character string specifying the name of the outcome variable used in the model. Must be provided if choose method = "multiplier".

time

A character string specifying the name of the time variable used in the model.

range

A numeric vector specifying the range or grid of time points interested. Defaut is the whole time points from the fitted model.

group_name

A character vector that specifies the names of grouping scalar variables to analyze. Default is NULL, representing the reference group if method = "cma". For method = "multiplier", NuULL is not allowed.

group_value

A numeric vector that specifies the corresponding values of the variables listed in group_name. Each entry in group_value should match the respective variable in group_name. Character/factor is not allowed. Default is NULL, representing the reference group if "cma". For method = "multiplier", NuULL is not allowed.

subject

A character string specifying the name of the subject-level random effect variable.

nboot

An integer specifying the number of bootstrap samples used to construct the confidence bands. Default is 10,000 for cma, 5000 for Multiplier Bootstrap.

method_SD

Method for SD estimation: "t" or "regular". Default is "t".

weights

Multiplier type: "rademacher", "gaussian", or "mammen". Default is "rademacher"."

For CMA bands (method = "cma"), please provide a pre-fitted functional regression model For Multiplier Bootstrap bands (method = "multiplier"), the original input data is required.

Value

A list containing:

yhat

Estimated mean function for the group of interest.

time

The time points used.

se_hat

Standard errors of the estimated means.

scb_low

Lower bound of the simultaneous confidence band.

scb_up

Upper bound of the simultaneous confidence band.

type

A character description of the output type.

Examples

# example using ccds data
data(ccds)
ccds_fpca <- prepare_ccds_fpca(ccds)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(subject, by = Phi1, bs="re") +
  s(subject, by = Phi2, bs="re")+
  s(subject, by = Phi3, bs="re") +
  s(subject, by = Phi4, bs="re"),
  method = "fREML", data = ccds_fpca, discrete = TRUE)

# CMA approach
results <- SCB_functional_outcome(data = ccds, object = fosr_mod, method = "cma", fitted = TRUE,
                                  est_mean = TRUE, outcome = "percent_change", time = "seconds",
                                  group_name = "use", group_value = 1, subject = "subject")

# multiplier bootstrap
results <- SCB_functional_outcome(data = ccds, object = fosr_mod, method = "multiplier", fitted = TRUE,
                       est_mean = TRUE, outcome = "percent_change",
                       time = "seconds", group_name = "use", group_value = 1, subject = "subject")
results <- SCB_functional_outcome(data = ccds, object = fosr_mod, method = "multiplier", fitted = TRUE,
                       est_mean = FALSE, outcome = "percent_change",
                       time = "seconds", group_name = "use", group_value = 1, subject = "subject")