Construct Simultaneous Confidence Bands for a Logistic Regression Model

This function fits a logistic regression model and constructs simultaneous confidence bands (SCB) using a non-parametric bootstrap method for the mean outcome of regression on a fixed test set design matrix

SCB_logistic_outcome(df_fit, model, grid_df, n_boot = 1000, alpha = 0.05)

Arguments

df_fit

A data frame containing the training design matrix used to fit the logistic model.

model

A character string representing the formula for the logistic model (e.g., "y ~ x1 + x2").

grid_df

A data frame containing the test set design matrix at which to evaluate the predictions and construct confidence bands.

n_boot

Number of bootstrap samples used in the non-parametric bootstrap procedure to generate the empirical distribution. Default is 1000.

alpha

Significance level for the confidence band (e.g., 0.05 for 95% confidence). Default is 0.05.

Value

A data frame with the following columns:

scb_low

Lower bound of the simultaneous confidence band.

Mean

Predicted mean response from the fitted model.

scb_up

Upper bound of the simultaneous confidence band.

grid_df

All columns from grid_df, representing the prediction grid.

Examples

set.seed(262)
x1 <- rnorm(100)
x2 <- rnorm(100)
epsilon <- rnorm(100,0,sqrt(2))
y <- -1 + x1 + 0.5 * x1^2 - 1.1 * x1^3 - 0.5 * x2 + 0.8 * x2^2 - 1.1 * x2^3 + epsilon
y <- expit(y)
df <- data.frame(x1 = x1, x2 = x2, y = y)
grid <- data.frame(x1 = seq(-1, 1, length.out = 100), x2 = seq(-1, 1, length.out = 100))
model <- "y ~ x1 + I(x1^2) + I(x1^3) + x2 + I(x2^2) + I(x2^3)"
results <- SCB_logistic_outcome(df_fit = df, model = model, grid_df = grid)