Historical and Future Summer Temperature Data in North America

A spatial dataset containing historical (1971-1999) and future (2041-2069) mean summer (June–August) surface temperatures over North America, used for evaluating increase of mean summer temperature between the 20th and 21st centuries in North America, and constructing simultaneous confidence bands via generalized least squares (GLS) modeling.

Usage

data(climate_data)

Format

A list with the following components:

Z

A list containing spatial data with three components: x (longitude), y (latitude), and z, a 3D array of observations with dimensions [lon, lat, n]. The first na slices of z come from mean summer temperature (June-August) in North America recorded from 1971 to 1999, and the last nb slices contain mean summer temperature from 2041 to 2069.

mask

A logical or numeric matrix of dimensions length(lon) × length(lat). Values are set to 1 for land and NA elsewhere based on the elevation matrix orog > 0.

X

A numeric design matrix with na + nb rows and 4 columns, constructed for generalized least squares (GLS) regression. The rows correspond to spatial replicates from na current years and nb future years. The columns are:

X1: Group indicator (0 for current years (1971-1999), 1 for future years (2041-2069))

X2: Intercept

X3: Centered time variable ta for current years (1971-1999) (0 for future years (2041-2069))

X4: Centered time variable tb for future years (2041-2069) (0 for current years (1971-1999))

correlation

A character string set to "corAR1", indicating that an autoregressive correlation structure of order 1 (AR(1)) is used for GLS fitting.

Source

Processed from data-raw/climate_data.R using the readr package.

Details

The data are arranged on a regular longitude–latitude grid, with spatial masking for land-only analysis. AR(1) correlation structure is assumed for statistical modeling.

References

Sommerfeld, M., Sain, S., & Schwartzman, A. (2018). Confidence regions for spatial excursion sets from repeated random field observations, with an application to climate. Journal of the American Statistical Association, 113(523), 1327–1340. doi:10.1080/01621459.2017.1356318