Construct CMA Confidence Intervals via Parametric Method

This function computes Correlation and Multiplicity Adjusted (CMA) confidence bands for a specified group in a functional outcome regression model using parameter simulations approach with Gaussian multiplier bootstrap.

Usage

cma(
  data_df,
  object,
  fitted = TRUE,
  alpha = 0.05,
  outcome,
  domain,
  subset = NULL,
  id,
  nboot = NULL
)

Arguments

data_df

A functional data frame that contain both name and values for variables including functional outcome, domain (e.g. time) and ID (e.g. subject names).

object

A fitted Function-on-Scalar Regression (FoSR) object (e.g., from mgcv::bam()/mgcv::gam()).

fitted

Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function

TRUE - Estimate the simultaneous confidence bands for fitted mean outcome function.

FALSE - estimate the simultaneous confidence bands for fitted parameter function.

Default is TRUE.

alpha

Significance level for SCB. Default is 0.05.

outcome

A character string specifying the name of the outcome variable used in the model.

domain

A character string specifying the name of the domain variable (e.g. time) used in the model.

subset

An atomic character vector (e.g., c("user = 1", "age = 30")) specified the target function for constructing the SCB. Each element must be of the form = , where is the name of a scalar grouping variable and is the desired value. Whitespace is ignored. Binary or categorical character variable should be transformed into numeric. Factor is not allowed here because if the input data contains factor variables, they will be automatically expanded into dummy (indicator) variables when constructing the design matrix, and the resulting variable names may differ from the original factor names. Default is NULL, representing the reference group.

id

A character string specifying the name of the ID variable.

nboot

An integer specifying the number of bootstrap samples used to construct the confidence bands. Default is 10,000.

Value

A list containing:

mu_hat

Estimated mean function for the group of interest.

domain

The domain used.

se_hat

Standard errors of the estimated means.

scb_low

Lower bound of the simultaneous confidence band.

scb_up

Upper bound of the simultaneous confidence band.

type

A character description of the output type.

Examples

# example using pupil data
library(mgcv)
#> Warning: package ‘mgcv’ was built under R version 4.4.3
#> Loading required package: nlme
#> Warning: package ‘nlme’ was built under R version 4.4.3
#> This is mgcv 1.9-3. For overview type 'help("mgcv-package")'.
data(pupil)
pupil_fpca <- prepare_pupil_fpca(pupil)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(id, by = Phi1, bs="re") +
  s(id, by = Phi2, bs="re")+
  s(id, by = Phi3, bs="re") +
  s(id, by = Phi4, bs="re"),
  method = "fREML", data = pupil_fpca, discrete = TRUE)

results <- cma(pupil_fpca, fosr_mod, fitted = TRUE, outcome = "percent_change",
               domain = "seconds", subset = c("use = 1"), id = "id")

results <- cma(pupil_fpca, fosr_mod, fitted = FALSE, outcome = "percent_change",
               domain = "seconds", subset = c("use = 1"), id = "id")