Functional Outcome Regression Prediction with Group-Specific Inference

This function is an internal function for constructing SCBs for functional data.

Usage

mean_response_predict(
  data_df,
  object,
  fitted = TRUE,
  time,
  range = NULL,
  group_name,
  group_value,
  subject = NULL
)

Arguments

data_df

A functional data frame. The input data must have column names, and should contain the functional outcome, time and subject

object

A fitted Function-on-Scalar Regression (FoSR) model object (e.g., from mgcv::bam()/mgcv::gam()).

fitted

Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function

TRUE - Estimate the simultaneous confidence bands for fitted mean outcome function.

FALSE - estimate the simultaneous confidence bands for fitted parameter function.

Default is TRUE.

time

A character string specifying the name of the time variable used in the model.

range

A numeric vector specifying the range or grid of time points interested. Defaut is the whole time points from the fitted model.

group_name

A character vector that specifies the names of grouping scalar variables to analyze. Default is NULL, representing the reference group.

group_value

A numeric vector that specifies the corresponding values of the variables listed in group_name. Each entry in group_value should match the respective variable in group_name. Character/factor is not allowed. Default is NULL, representing the reference group.

subject

A optional character string specifying the name of the subject-level random effect variable, if included in model fitting.

Value

A list containing the following elements:

s_pred

Numeric vector of sorted unique time points used for prediction

pred_df

Data frame with prediction results, containing:

mean: Predicted mean values

se: Standard errors

lpmat

Linear predictor matrix (design matrix) used for confidence interval calculations

mod_coef

Vector of model coefficients for selected group

mod_cov

Variance-covariance matrix corresponding to the selected group coefficients

Examples

library(mgcv)
data(pupil)
pupil_fpca <- prepare_pupil_fpca(pupil)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(id, by = Phi1, bs="re") +
  s(id, by = Phi2, bs="re")+
  s(id, by = Phi3, bs="re") +
  s(id, by = Phi4, bs="re"),
  method = "fREML", data = pupil_fpca, discrete = TRUE)

results <- mean_response_predict(pupil_fpca, fosr_mod, fitted = TRUE, time = "seconds",
group_name = "use", group_value = 1, subject = "id")

results <- mean_response_predict(pupil_fpca, fosr_mod, fitted = FALSE, time = "seconds",
group_name = "use", group_value = 1, subject = "id")