Construct Inverse Confidence Sets from Simultaneous Confidence Bands

This function constructs inverse confidence sets (CS) from simultaneous confidence bands (SCB), allowing visualization and containment check of level sets of true or estimated functions.

scb_to_cs(
  scb_up,
  scb_low,
  levels,
  true_mean = NULL,
  est_mean = NULL,
  x1,
  x2,
  type = "upper",
  return_contain_only = F,
  return_plot = F,
  xlab = NULL,
  ylab = NULL
)

Arguments

scb_up

A matrix or array containing the upper simultaneous confidence interval. Must have the same dimensions as defined by the grid from x1 and x2.

scb_low

A matrix or array containing the lower bounds of the simultaneous confidence bands. Must have the same dimensions as defined by the grid from x1 and x2.

levels

A list of scalers for different levels or matrix containing interval sets to construct the confidence sets.

true_mean

Optional matrix of the true mean function (for simulation study or evaluation purposes). Should have the same dimension as scb_up.

est_mean

Optional matrix of the estimated mean function, used for plotting if true_mean is not available. Should have the same dimension as scb_up.

x1

A numeric vector of coordinates for the first dimension (e.g., time or x-axis).

x2

A numeric vector of coordinates for the second dimension (e.g., space or y-axis).

type

A character string specifying the type of inverse set to construct if levels are not a matrix. Choices are "upper", "lower", "two-sided" or "interval". Notice that "two-sided" type is not available for plotting (return_plot = TRUE).

return_contain_only

Logical. If TRUE, only return a matrix/logical map indicating whether the level set is contained.

return_plot

Logical. If TRUE, return a ggplot object for visualization.

xlab

A character for the name of the x axis for the returned ggplot object.

ylab

A character for the name of the y axis for the returned ggplot object.

Value

A list containing the following components:

levels

A vector (or matrix) of threshold levels used to define the confidence sets.

U_in

(Optional) A list of logical matrices indicating whether each point is within the conservative inner confidence set for each level. Returned only when return_contain_only = FALSE and type != "two-sided".

U_out

(Optional) A list of logical matrices indicating whether each point is within the liberal outer confidence set for each level. Returned only when return_contain_only = FALSE and type != "two-sided".

L_out

(Two-sided only) A list of logical matrices indicating lower bound containment (for type = "two-sided").

U_out

(Two-sided only) A list of logical matrices indicating upper bound containment (for type = "two-sided").

contain_individual

A logical vector indicating whether the true mean is fully contained within each level's confidence set. Returned only if true_mean is provided.

contain_all

A single logical value indicating whether the true mean is contained in all levels' confidence sets.

plot_cs

(Optional) A list of ggplot2 objects for visualizing the confidence sets, returned when return_plot = TRUE. Includes both a combined plot and individual plots per level.

Examples

set.seed(262)
x1 <- rnorm(100)
x2 <- rnorm(100)
y <- -1 + x1 + 0.5 * x1^2 - 1.1 * x1^3 - 0.5 * x2 + 0.8 * x2^2 - 1.1 * x2^3 - rnorm(100,0,sqrt(2))
df <- data.frame(x1 = x1, x2 = x2, y = y)
grid <- data.frame(x1 = seq(-1, 1, length.out = 100), x2 = seq(-1, 1, length.out = 100))
model <- "y ~ x1 + I(x1^2) + I(x1^3) + x2 + I(x2^2) + I(x2^3)"
result <- SCB_linear_outcome(df_fit = df, model = model, grid_df = grid)
scb_to_cs(result$UpperBound, result$LowerBound, c(-1, -0.5, 0.5, 1),
x1 = grid$x1, x2 = grid$x2, est_mean = results$Mean)
#> $levels
#> [1] -1.0 -0.5  0.5  1.0
#> 
#> $U_in
#> $U_in[[1]]
#> logical(0)
#> 
#> $U_in[[2]]
#> logical(0)
#> 
#> $U_in[[3]]
#> logical(0)
#> 
#> $U_in[[4]]
#> logical(0)
#> 
#> 
#> $U_out
#> $U_out[[1]]
#> logical(0)
#> 
#> $U_out[[2]]
#> logical(0)
#> 
#> $U_out[[3]]
#> logical(0)
#> 
#> $U_out[[4]]
#> logical(0)
#> 
#> 
#> $contain_individual
#> NULL
#> 
#> $contain_all
#> [1] TRUE
#> 
#> $plot_cs
#> NULL
#>